refactor: 역할별 파일 분리
